Introducing DAO Interface.

+++ src/DAO/AreaDAO.java
@@ -11,7 +11,7 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
-public class AreaDAO {
+public class AreaDAO implements DAO {
 
     Connection conn;
     PreparedStatement pstm;
@@ -30,7 +30,7 @@ public class AreaDAO {
 
             while (rs.next()) {
                 AreaDTO objAreaDTO = new AreaDTO();
-                objAreaDTO.setIdArea(rs.getInt("id_area"));
+                objAreaDTO.setId(rs.getInt("id_area"));
                 objAreaDTO.setNomeArea(rs.getString("nome_area"));
 
                 listaAreas.add(objAreaDTO);
@@ -42,52 +42,70 @@ public class AreaDAO {
         return listaAreas;
     }
 
-    public void cadastrarArea(AreaDTO objAreaDTO) {
-        String sql = "INSERT INTO areas (nome_area) VALUES (?);";
+    @Override
+    public void insert(Object obj) {
+        if (obj instanceof AreaDTO) {
+            AreaDTO objAreaDTO = (AreaDTO) obj;
+            String sql = "INSERT INTO areas (nome_area) VALUES (?);";
 
-        conn = new ConexaoDAO().conectaBD();
+            conn = new ConexaoDAO().conectaBD();
 
-        try {
-            pstm = conn.prepareStatement(sql);
-            pstm.setString(1, objAreaDTO.getNomeArea());
+            try {
+                pstm = conn.prepareStatement(sql);
+                pstm.setString(1, objAreaDTO.getNomeArea());
 
-            pstm.execute();
-            pstm.close();
-        } catch (Exception e) {
-            Alerts.showAlert("Error", null,"AreaDAO Cadastrar" + e.getMessage(), Alert.AlertType.ERROR);
+                pstm.execute();
+                pstm.close();
+            } catch (Exception e) {
+                Alerts.showAlert("Error", null,"AreaDAO Cadastrar" + e.getMessage(), Alert.AlertType.ERROR);
+            }
+        } else {
+            Alerts.showAlert("Error", null,"Objeto Inválido", Alert.AlertType.ERROR);
         }
     }
 
-    public void updateArea(AreaDTO objAreaDTO) {
-        String sql = "UPDATE areas SET nome_area = ? WHERE id_area = ?;";
+    @Override
+    public void update(Object obj) {
+        if (obj instanceof AreaDTO) {
+            AreaDTO objAreaDTO = (AreaDTO) obj;
+            String sql = "UPDATE areas SET nome_area = ? WHERE id_area = ?;";
 
-        conn = new ConexaoDAO().conectaBD();
+            conn = new ConexaoDAO().conectaBD();
 
-        try {
-            pstm = conn.prepareStatement(sql);
-            pstm.setString(1, objAreaDTO.getNomeArea());
-            pstm.setInt(2, objAreaDTO.getIdArea());
+            try {
+                pstm = conn.prepareStatement(sql);
+                pstm.setString(1, objAreaDTO.getNomeArea());
+                pstm.setInt(2, objAreaDTO.getId());
 
-            pstm.execute();
-            pstm.close();
-        } catch (Exception e) {
-            Alerts.showAlert("Error", null,"AreaDAO Update" + e.getMessage(), Alert.AlertType.ERROR);
+                pstm.execute();
+                pstm.close();
+            } catch (Exception e) {
+                Alerts.showAlert("Error", null,"AreaDAO Update" + e.getMessage(), Alert.AlertType.ERROR);
+            }
+        } else {
+            Alerts.showAlert("Error", null,"Objeto Inválido", Alert.AlertType.ERROR);
         }
     }
 
-    public void deleteArea(AreaDTO objAreaDTO) {
-        String sql = "DELETE FROM areas WHERE id_area = ?;";
+    @Override
+    public void delete(Object obj) {
+        if (obj instanceof AreaDTO) {
+            AreaDTO objAreaDTO = (AreaDTO) obj;
+            String sql = "DELETE FROM areas WHERE id_area = ?;";
 
-        conn = new ConexaoDAO().conectaBD();
+            conn = new ConexaoDAO().conectaBD();
 
-        try {
-            pstm = conn.prepareStatement(sql);
-            pstm.setInt(1, objAreaDTO.getIdArea());
+            try {
+                pstm = conn.prepareStatement(sql);
+                pstm.setInt(1, objAreaDTO.getId());
 
-            pstm.execute();
-            pstm.close();
-        } catch (SQLException e) {
-            Alerts.showAlert("Error", null,"AreaDAO Delete" + e.getMessage(), Alert.AlertType.ERROR);
+                pstm.execute();
+                pstm.close();
+            } catch (Exception e) {
+                Alerts.showAlert("Error", null,"AreaDAO Delete" + e.getMessage(), Alert.AlertType.ERROR);
+            }
+        } else {
+            Alerts.showAlert("Error", null,"Objeto Inválido", Alert.AlertType.ERROR);
         }
     }
 }
